+++ packages/env-variables/src/api.ts
@@ -33,6 +33,8 @@ export const privateEnvVariablesSchema = z.object({
 	OPEN_AI_KEY: z.string().optional(),
 	GOOGLE_TRANSLATE_API_KEY: z.string().optional(),
 	SERVER_SENTRY_DSN: z.string().optional().describe("DSN for Sentry (on the server)"),
+	ALGOLIA_ADMIN: z.string(),
+	ALGOLIA_APPLICATION: z.string(),
 	// prefixed with INLANG_ because github doesn't allow env vars with GITHUB_ in ci/cd.
 	INLANG_GITHUB_APP_CLIENT_SECRET: z.string(),
 })