+++ packages/versioned-interfaces/module/src/interface.test-d.ts
@@ -1,0 +1,5 @@
+import type { InlangModule } from "./interface.js"
+
+// test: an inlang module should have a settingsSchema
+const module: InlangModule["default"] = {} as any
+module.settingsSchema