+++ packages/ide-extension/CHANGELOG.md
@@ -1,5 +1,11 @@
 # inlang-vs-code-extension
 
+## 1.14.0
+
+### Minor Changes
+
+- 77ed7a85c: update deps
+
 ## 1.13.0
 
 ### Minor Changes