+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/sdk
 
+## 0.16.0
+
+### Minor Changes
+
+- 2f924df32: added Modulesettings validation via the Typebox JSON Schema Validation. This ensure that users can exclusively use module settings when there are given by the moduel
+
+### Patch Changes
+
+- Updated dependencies [2f924df32]
+  - @inlang/message-lint-rule@1.4.0
+  - @inlang/plugin@2.2.0
+
 ## 0.15.0
 
 ### Minor Changes